+++ app/src/main/java/com/example/myapplication/models/User.java
@@ -2,13 +2,14 @@ package com.example.myapplication.models;
 
 public class User {
 
-    private String user_id;
+//    private String user_id;
     private long phone_number;
     private String email;
     private String username;
 
-    public User(String user_id, long phone_number, String email, String username) {
-        this.user_id = user_id;
+//    public User(String user_id, long phone_number, String email, String username) {
+    public User( long phone_number, String email, String username) {
+//        this.user_id = user_id;
         this.phone_number = phone_number;
         this.email = email;
         this.username = username;
@@ -18,13 +19,13 @@ public class User {
 
     }
 
-    public String getUser_id() {
-        return user_id;
-    }
-
-    public void setUser_id(String user_id) {
-        this.user_id = user_id;
-    }
+//    public String getUser_id() {
+//        return user_id;
+//    }
+//
+//    public void setUser_id(String user_id) {
+//        this.user_id = user_id;
+//    }
 
     public long getPhone_number() {
         return phone_number;
@@ -53,7 +54,7 @@ public class User {
     @Override
     public String toString() {
         return "User{" +
-                "user_id='" + user_id + '\'' +
+//                "user_id='" + user_id + '\'' +
                 ", phone_number='" + phone_number + '\'' +
                 ", email='" + email + '\'' +
                 ", username='" + username + '\'' +
