+++ app/src/main/java/com/example/myapplication/user_profile/EditProfileFragment.java
@@ -113,7 +113,7 @@ public class EditProfileFragment extends Fragment implements View.OnClickListene
 //                Log.d(TAG, "onDataChange: Current Username: " + user.getUsername());
 
                 if (!mUserSettings.getUser().getUsername().equals(userName)) {
-                    checkIfUsernameExists(userName);
+                    firebaseMethods.checkIfUsernameExists(userName);
                 }
             }
 
@@ -123,51 +123,6 @@ public class EditProfileFragment extends Fragment implements View.OnClickListene
             }
         });
     }
-
-    /**
-     * Check if @param username already exists
-     *
-     * @param userName of the User from Firebase database
-     */
-    private void checkIfUsernameExists(final String userName) {
-        Log.d(TAG, "checkIfUsernameExists: checking if " + userName + " already exists");
-
-        DatabaseReference reference = mFirebaseDatabase.getReference();
-
-        Query query = reference
-                .child(getString(R.string.dbname_users))
-                .orderByChild(getString(R.string.field_username))
-                .equalTo(userName);
-
-
-        query.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                if (!dataSnapshot.exists()) {
-                    //add username
-                    firebaseMethods.updateUsername(userName);
-                    Toast.makeText(getActivity(), "saved username", Toast.LENGTH_SHORT).show();
-
-                }
-
-
-                for (DataSnapshot ds : dataSnapshot.getChildren()) {
-                    if (ds.exists()) {
-                        Log.d(TAG, "onDataChange: Found a Match: " + ds.getValue(User.class).getUsername());
-                        Toast.makeText(getActivity(), "That username already exists", Toast.LENGTH_SHORT).show();
-
-                    }
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError databaseError) {
-
-            }
-        });
-
-    }
-
 
     private void setProfileWidgets(UserSettings userSettings) {
         Log.d(TAG, "setProfileWidgets: setting widgets with data, retrieving from database: " +
@@ -250,7 +205,7 @@ public class EditProfileFragment extends Fragment implements View.OnClickListene
 
                 // retrive user information from the database
 //                (firebaseMethods.getUserSettings(dataSnapshot));
-                setProfileWidgets(getUserSettings(dataSnapshot));
+                setProfileWidgets(firebaseMethods.getUserSettings(dataSnapshot));
 
                 //retrive images for the user in question
 
@@ -262,117 +217,6 @@ public class EditProfileFragment extends Fragment implements View.OnClickListene
             }
         });
 
-    }
-
-    /**
-     * Retrieves the account settings for the User currently logged in
-     * Database:user_account_settings node
-     *
-     * @param dataSnapshot represent the data from database
-     * @return the User Account Settings
-     */
-    private UserSettings getUserSettings(DataSnapshot dataSnapshot) {
-        Log.d(TAG, "getUserAccountSettings: retrieving user account settings from database");
-
-        UserAccountSettings settings = new UserAccountSettings();
-        User user = new User();
-        String userID = mAuth.getCurrentUser().getUid();
-
-        for (DataSnapshot ds : dataSnapshot.getChildren()) {
-
-            //User Account Settings Node
-            if (ds.getKey().equals(mContext.getString(R.string.dbname_user_account_settings))) {
-                Log.d(TAG, "getUserAccountSettings: dataSnapshot" + ds);
-                try {
-                    settings.setUsername(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getUsername()
-                    );
-
-                    settings.setDisplay_name(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getDisplay_name()
-                    );
-
-                    settings.setDescription(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getDescription()
-                    );
-
-                    settings.setWebsite(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getWebsite()
-                    );
-
-                    settings.setFollowers(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getFollowers()
-                    );
-
-                    settings.setFollowing(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getFollowing()
-                    );
-
-                    settings.setPosts(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getPosts()
-                    );
-
-                    settings.setProfile_photo(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getProfile_photo()
-                    );
-
-                    Log.d(TAG, "getUserAccountSettings: retrieve user account settings information: " + settings.toString());
-                } catch (NullPointerException e) {
-                    Log.d(TAG, "getUserAccountSettings: NullPointerException: " + e.getMessage());
-                }
-            }
-
-            //Users Node
-            if (ds.getKey().equals(mContext.getString(R.string.dbname_users))) {
-                Log.d(TAG, "getUserAccountSettings: dataSnapshot" + ds);
-
-                user.setUser_id(
-                        ds.child(userID)
-                                .getValue(User.class)
-                                .getUser_id()
-                );
-
-                user.setUsername(
-                        ds.child(userID)
-                                .getValue(User.class)
-                                .getUsername()
-                );
-
-                user.setEmail(
-                        ds.child(userID)
-                                .getValue(User.class)
-                                .getEmail()
-                );
-
-                user.setPhone_number(
-                        ds.child(userID)
-                                .getValue(User.class)
-                                .getPhone_number()
-                );
-
-                Log.d(TAG, "getUserInformation: retrieve user  information: " + user.toString());
-
-            }
-
-        }
-
-        return new UserSettings(user, settings);
     }
 
 }
