+++ app/src/main/java/com/example/myapplication/user_profile/UserProfileFragment.java
@@ -149,7 +149,7 @@ public class UserProfileFragment extends Fragment implements View.OnClickListene
             public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
 
                 //retrive user information from the database
-                setProfileWidgets(getUserSettings(dataSnapshot));
+                setProfileWidgets(firebaseMethods.getUserSettings(dataSnapshot));
 
                 //retrive images for the user in question
 
@@ -201,117 +201,6 @@ public class UserProfileFragment extends Fragment implements View.OnClickListene
                 break;
         }
 
-    }
-
-    /**
-     * Retrieves the account settings for the User currently logged in
-     * Database:user_account_settings node
-     *
-     * @param dataSnapshot represent the data from database
-     * @return the User Account Settings
-     */
-    private UserSettings getUserSettings(DataSnapshot dataSnapshot) {
-        Log.d(TAG, "getUserAccountSettings: retrieving user account settings from database");
-
-        UserAccountSettings settings = new UserAccountSettings();
-        User user = new User();
-        String userID = mAuth.getCurrentUser().getUid();
-
-        for (DataSnapshot ds : dataSnapshot.getChildren()) {
-
-            //User Account Settings Node
-            if (ds.getKey().equals(mContext.getString(R.string.dbname_user_account_settings))) {
-                Log.d(TAG, "getUserAccountSettings: dataSnapshot" + ds);
-                try {
-                    settings.setUsername(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getUsername()
-                    );
-
-                    settings.setDisplay_name(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getDisplay_name()
-                    );
-
-                    settings.setDescription(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getDescription()
-                    );
-
-                    settings.setWebsite(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getWebsite()
-                    );
-
-                    settings.setFollowers(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getFollowers()
-                    );
-
-                    settings.setFollowing(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getFollowing()
-                    );
-
-                    settings.setPosts(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getPosts()
-                    );
-
-                    settings.setProfile_photo(
-                            ds.child(userID)
-                                    .getValue(UserAccountSettings.class)
-                                    .getProfile_photo()
-                    );
-
-                    Log.d(TAG, "getUserAccountSettings: retrieve user account settings information: " + settings.toString());
-                } catch (NullPointerException e) {
-                    Log.d(TAG, "getUserAccountSettings: NullPointerException: " + e.getMessage());
-                }
-            }
-
-            //Users Node
-            if (ds.getKey().equals(mContext.getString(R.string.dbname_users))) {
-                Log.d(TAG, "getUserAccountSettings: dataSnapshot" + ds);
-
-                user.setUser_id(
-                        ds.child(userID)
-                                .getValue(User.class)
-                                .getUser_id()
-                );
-
-                user.setUsername(
-                        ds.child(userID)
-                                .getValue(User.class)
-                                .getUsername()
-                );
-
-                user.setEmail(
-                        ds.child(userID)
-                                .getValue(User.class)
-                                .getEmail()
-                );
-
-                user.setPhone_number(
-                        ds.child(userID)
-                                .getValue(User.class)
-                                .getPhone_number()
-                );
-
-                Log.d(TAG, "getUserInformation: retrieve user  information: " + user.toString());
-
-            }
-
-        }
-
-        return new UserSettings(user, settings);
     }
 
     /**
