+++ app/src/main/java/com/example/myapplication/utility_classes/FirebaseMethods.java
@@ -3,6 +3,7 @@ package com.example.myapplication.utility_classes;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.util.Log;
+import android.widget.Toast;
 
 import com.example.myapplication.R;
 import com.example.myapplication.models.User;
@@ -12,8 +13,11 @@ import com.example.myapplication.user_profile.UserProfileActivity;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.Query;
+import com.google.firebase.database.ValueEventListener;
 
 /**
  * File created by tcarau18
@@ -65,7 +69,7 @@ public class FirebaseMethods {
 
     public void addNewUser(String email, String username, String description, String website, String profile_photo) {
 
-        User user = new User(email, 1, email, StringManipulation.condenseUserName(username));
+        User user = new User( 1, email, StringManipulation.condenseUserName(username));
 
         myRef.child(mContext.getString(R.string.dbname_users))
                 .child(userUID)
@@ -120,7 +124,155 @@ public class FirebaseMethods {
                 .setValue(username);
     }
 
+    /**
+     * Check if @param username already exists
+     *
+     * @param userName of the User from Firebase database
+     */
+    public void checkIfUsernameExists(final String userName) {
+        Log.d(TAG, "checkIfUsernameExists: checking if " + userName + " already exists");
+
+        DatabaseReference reference = mFirebaseDatabase.getReference();
+
+        Query query = reference
+                .child(String.valueOf(R.string.dbname_users))
+                .orderByChild(String.valueOf(R.string.field_username))
+                .equalTo(userName);
+
+
+        query.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                if (!dataSnapshot.exists()) {
+                    //add username
+                    updateUsername(userName);
+                    Toast.makeText(mContext, "saved username", Toast.LENGTH_SHORT).show();
+
+                }
+
+
+                for (DataSnapshot ds : dataSnapshot.getChildren()) {
+                    if (ds.exists()) {
+                        Log.d(TAG, "onDataChange: Found a Match: " + ds.getValue(User.class).getUsername());
+                        Toast.makeText(mContext, "That username already exists", Toast.LENGTH_SHORT).show();
+
+                    }
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+
+            }
+        });
+
+    }
 
 //    static/lambda/((UserProfileActivity.class).getctvoty)
 
+    /**
+     * Retrieves the account settings for the User currently logged in
+     * Database:user_account_settings node
+     *
+     * @param dataSnapshot represent the data from database
+     * @return the User Account Settings
+     */
+    public UserSettings getUserSettings(DataSnapshot dataSnapshot) {
+        Log.d(TAG, "getUserAccountSettings: retrieving user account settings from database");
+
+        UserAccountSettings settings = new UserAccountSettings();
+        User user = new User();
+        String userID = mAuth.getCurrentUser().getUid();
+
+        for (DataSnapshot ds : dataSnapshot.getChildren()) {
+
+            //User Account Settings Node
+            if (ds.getKey().equals(mContext.getString(R.string.dbname_user_account_settings))) {
+                Log.d(TAG, "getUserAccountSettings: dataSnapshot" + ds);
+                try {
+                    settings.setUsername(
+                            ds.child(userID)
+                                    .getValue(UserAccountSettings.class)
+                                    .getUsername()
+                    );
+
+                    settings.setDisplay_name(
+                            ds.child(userID)
+                                    .getValue(UserAccountSettings.class)
+                                    .getDisplay_name()
+                    );
+
+                    settings.setDescription(
+                            ds.child(userID)
+                                    .getValue(UserAccountSettings.class)
+                                    .getDescription()
+                    );
+
+                    settings.setWebsite(
+                            ds.child(userID)
+                                    .getValue(UserAccountSettings.class)
+                                    .getWebsite()
+                    );
+
+                    settings.setFollowers(
+                            ds.child(userID)
+                                    .getValue(UserAccountSettings.class)
+                                    .getFollowers()
+                    );
+
+                    settings.setFollowing(
+                            ds.child(userID)
+                                    .getValue(UserAccountSettings.class)
+                                    .getFollowing()
+                    );
+
+                    settings.setPosts(
+                            ds.child(userID)
+                                    .getValue(UserAccountSettings.class)
+                                    .getPosts()
+                    );
+
+                    settings.setProfile_photo(
+                            ds.child(userID)
+                                    .getValue(UserAccountSettings.class)
+                                    .getProfile_photo()
+                    );
+
+                    Log.d(TAG, "getUserAccountSettings: retrieve user account settings information: " + settings.toString());
+                } catch (NullPointerException e) {
+                    Log.d(TAG, "getUserAccountSettings: NullPointerException: " + e.getMessage());
+                }
+            }
+
+            //Users Node
+            if (ds.getKey().equals(mContext.getString(R.string.dbname_users))) {
+                Log.d(TAG, "getUserAccountSettings: dataSnapshot" + ds);
+
+
+                user.setUsername(
+                        ds.child(userID)
+                                .getValue(User.class)
+                                .getUsername()
+                );
+
+                user.setEmail(
+                        ds.child(userID)
+                                .getValue(User.class)
+                                .getEmail()
+                );
+
+                user.setPhone_number(
+                        ds.child(userID)
+                                .getValue(User.class)
+                                .getPhone_number()
+                );
+
+                Log.d(TAG, "getUserInformation: retrieve user  information: " + user.toString());
+
+            }
+
+        }
+
+        return new UserSettings(user, settings);
+    }
 }
